update course to work

+++ controllers/coursesController.js
@@ -2,37 +2,46 @@ const responseHelper = require('../helpers/responseHelper');
 const courseService = require('../services/courseService');
 const copyObjectHelper = require('../helpers/propertyCopyHelper');
 const db = require('../database/database');
-
-// TODO: remove error as soon as they are implemented.
-
-const throwError = () => {
-  throw false;
-};
+const { checkCourseEditAuthorization } = require('../helpers/checkAuthorizationHelper');
 
 exports.getCourses = async (req, res, next) => {
-  const directorOfStudiesId = req.token.userId;
+  const directorOfStudiesId = req.token.directorOfStudies_id;
   try {
-    // find all to current dos
-    const courses = await courseService.findAll(true, true, false, directorOfStudiesId);
-    responseHelper(res, 200, { courses });
+    const Courses = await courseService.findAll(true, true, false, directorOfStudiesId);
+    responseHelper(res, 200, 'Successful', { Courses });
   } catch (error) {
     return next(error);
   }
 };
 
+function getDirectorsOfStudiesArray(dosArray, currentDoS) {
+  if (!dosArray) {
+    dosArray = [];
+  }
+  if (!dosArray.includes(currentDoS)) {
+    dosArray.push(currentDoS);
+  }
+  return dosArray;
+}
+
 exports.postCourses = async (req, res, next) => {
-  const directorOfStudiesId = req.token.userId;
+  const directorOfStudiesId = req.token.directorOfStudies_id;
 
-  let transaction = await db.sequelize.transaction();
+  const transaction = await db.sequelize.transaction();
   try {
-    let courseToCreate = copyObjectHelper(req.body, ['name', 'majorSubject_id', 'directorOfStudy_ids', 'semesters']);
+    const courseToCreate = copyObjectHelper(req.body, [
+      'name',
+      'majorSubject_id',
+      'google_calendar_id',
+      'directorOfStudies_ids',
+      'Semesters',
+    ]);
 
-    if (!courseToCreate.directorOfStudy_ids) {
-      courseToCreate.directorOfStudy_ids = [];
-    }
-    courseToCreate.directorOfStudy_ids.push(directorOfStudiesId);
-
-    let createdCourse = await courseService.createCourse(transaction, { ...courseToCreate }, true, true);
+    courseToCreate.directorOfStudies_ids = getDirectorsOfStudiesArray(
+      courseToCreate.directorOfStudies_ids,
+      directorOfStudiesId
+    );
+    const createdCourse = await courseService.createCourse(transaction, { ...courseToCreate });
 
     transaction.commit();
 
@@ -44,25 +53,33 @@ exports.postCourses = async (req, res, next) => {
 };
 
 exports.putCourses = async (req, res, next) => {
-  // TODO: security: only if courseId belongs to DoS
-  const courseId = req.query.courseId;
-  const directorOfStudiesId = req.token.userId;
+  const course_id = req.query.courseId;
+  const directorOfStudiesId = req.token.directorOfStudies_id;
 
-  let transaction = await db.sequelize.transaction();
+  const transaction = await db.sequelize.transaction();
 
   try {
-    let courseToUpdate = copyObjectHelper(req.body, ['name', 'majorSubject_id', 'directorOfStudy_ids']);
+    if (await checkCourseEditAuthorization(directorOfStudiesId, course_id)) {
+      const courseToUpdate = copyObjectHelper(req.body, [
+        'name',
+        'majorSubject_id',
+        'directorOfStudies_ids',
+        'google_calendar_id',
+      ]);
 
-    if (!courseToUpdate.directorOfStudy_ids) {
-      courseToUpdate.directorOfStudy_ids = [];
+      courseToUpdate.directorOfStudies_ids = getDirectorsOfStudiesArray(
+        courseToUpdate.directorOfStudies_ids,
+        directorOfStudiesId
+      );
+
+      const updatedCourse = await courseService.updateCourse(transaction, { course_id, ...courseToUpdate });
+
+      transaction.commit();
+
+      return responseHelper(res, 200, 'Successfully updated.', updatedCourse);
+    } else {
+      return responseHelper(res, 400, 'You are not authorized to update this course.');
     }
-    courseToUpdate.directorOfStudy_ids.push(directorOfStudiesId);
-
-    let updatedCourse = await courseService.updateCourse(transaction, { courseId, ...courseToUpdate });
-
-    transaction.commit();
-
-    return responseHelper(res, 200, 'Successfully updated.', updatedCourse);
   } catch (error) {
     transaction.rollback();
     return next(error);
@@ -70,18 +87,21 @@ exports.putCourses = async (req, res, next) => {
 };
 
 exports.deleteCourses = async (req, res, next) => {
-  // TODO: security: only if courseId belongs to DoS
   const courseId = req.query.courseId;
-  const directorOfStudiesId = req.token.userId;
+  const directorOfStudiesId = req.token.directorOfStudies_id;
 
-  let transaction = await db.sequelize.transaction();
+  const transaction = await db.sequelize.transaction();
 
   try {
-    let deletedCourse = await courseService.deleteCourse(transaction, courseId);
+    if (await checkCourseEditAuthorization(directorOfStudiesId, courseId)) {
+      const deletedCourse = await courseService.deleteCourse(transaction, courseId);
 
-    transaction.commit();
+      transaction.commit();
 
-    return responseHelper(res, 200, 'Successfully deleted.', deletedCourse);
+      return responseHelper(res, 200, 'Successfully deleted.', deletedCourse);
+    } else {
+      return responseHelper(res, 400, 'You are not authorized to delete this course.');
+    }
   } catch (error) {
     transaction.rollback();
     return next(error);
